% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StructuralDecompose.R
\name{LevelCheck}
\alias{LevelCheck}
\title{minimum level length checks}
\usage{
LevelCheck(timeseries, level_length = 10, Breakpoints = c(), breaks)
}
\arguments{
\item{timeseries}{Given time series}

\item{level_length}{Mean distance between two levels}

\item{Breakpoints}{Breakpoints identified}

\item{breaks}{breakpoints returned}
}
\value{
The series cleaned with the minimum level check
}
\description{
minimum level length checks
}
