% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StructuralDecompose.R
\name{StructuralDecompose}
\alias{StructuralDecompose}
\title{Main decomposition algorithm}
\usage{
StructuralDecompose(
  Data,
  frequency = 12,
  break_algorithm = "strucchange",
  smoothening_algorithm = "lowess",
  break_level = 0.05,
  median_level = 0.5,
  mean_level = 0.5,
  level_length = 0.5,
  conf_level = 0.5
)
}
\arguments{
\item{Data}{Time series required}

\item{frequency}{Frequency of the tine series}

\item{break_algorithm}{breakpoints algorithm used. Defaults to strucchange}

\item{smoothening_algorithm}{Smoothing algorithm used. Defaults to lowess}

\item{break_level}{Break level for the breakpoints algorithm}

\item{median_level}{Average median distance between two level}

\item{mean_level}{Average mean distance between a group of points near breakpoints}

\item{level_length}{Minimum number of points required to determine a level}

\item{conf_level}{Confidence interval used for Anomaly detection}
}
\value{
The decomposed time series along with a host of other metrics
}
\description{
Main decomposition algorithm
}
