% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StructuralDecompose.R
\name{LevelCheck}
\alias{LevelCheck}
\title{Minimum level length checks}
\usage{
LevelCheck(timeseries, level_length = 10, breaks)
}
\arguments{
\item{timeseries}{Given time series}

\item{level_length}{Mean distance between two levels}

\item{breaks}{breakpoints returned}
}
\value{
The series cleaned with the minimum level check
}
\description{
Minimum level length checks
}
\examples{
LevelCheck(timeseries = StructuralDecompose::Nile_dataset[,1], breaks = c(1,4,5))

LevelCheck(timeseries = runif(n = 50, min = 1, max = 10), breaks = c(1,4,5))
}
