% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StructuralDecompose.R
\name{MeanCleaning}
\alias{MeanCleaning}
\title{Mean level checks}
\usage{
MeanCleaning(
  timeseries,
  mean_level = 0.5,
  Breakpoints = c(),
  breaks,
  frequency = 52
)
}
\arguments{
\item{timeseries}{Given time series}

\item{mean_level}{Mean distance between two levels}

\item{Breakpoints}{Breakpoints identified}

\item{breaks}{breakpoints returned}

\item{frequency}{Tiemseries frequency, defaults to 12 points}
}
\value{
The series cleaned with the mean check
}
\description{
Mean level checks
}
