% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StructuralDecompose.R
\name{MedianCleaning}
\alias{MedianCleaning}
\title{Median level checks}
\usage{
MedianCleaning(
  timeseries,
  median_level = 0.5,
  Breakpoints = c(),
  breaks,
  frequency = 52
)
}
\arguments{
\item{timeseries}{Given time series}

\item{median_level}{Median distance between two levels}

\item{Breakpoints}{Breakpoints identified}

\item{breaks}{Breaks identified}

\item{frequency}{Tiemseries frequency, defaults to 12 points}
}
\value{
The series cleaned with the median check
}
\description{
Median level checks
}
