% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StructuralDecompose.R
\name{Smoothing}
\alias{Smoothing}
\title{Smoothening of the time series}
\usage{
Smoothing(
  timeseries,
  frequency = 52,
  smoothening_algorithm = "lowess",
  breaks = c(0)
)
}
\arguments{
\item{timeseries}{Given time series}

\item{frequency}{Tiemseries frequency, defaults to 12 points}

\item{smoothening_algorithm}{Smoothening algorithm required}

\item{breaks}{Structural Breaks coming in from the break point algorithm}

\item{lowess}{Lowess smoothener}
}
\value{
The smoothened time series
}
\description{
Smoothening of the time series
}
